search/job_card_extractor.py
"""
Extract job cards from LinkedIn search results.

When you load a LinkedIn jobs search page it contains a list of job
cards, each representing a single job listing. These cards include
links to the job detail page as well as metadata such as the job title,
company, location and whether it supports Easy Apply. This module
provides a placeholder function for parsing those cards using
Playwright. In a full implementation you would need to inspect the
page structure and select the appropriate elements.
"""
from __future__ import annotations

from typing import List, Dict, Any

from jobagent.browser.playwright_driver import BrowserDriver


def extract_job_cards(driver: BrowserDriver) -> List[Dict[str, Any]]:
    """Extract job cards from the current search results page.

    Args:
        driver: An instance of ``BrowserDriver`` positioned on a LinkedIn
            job search page.

    Returns:
        A list of dictionaries. Each dictionary should have at minimum a
        ``url`` key pointing to the job detail page and a boolean
        ``easy_apply`` flag indicating whether the job supports Easy
        Apply. Additional metadata such as ``title``, ``company`` and
        ``location`` can also be included.
    """
    # TODO: implement job card parsing via Playwright queries
    return []