requirements.txt
playwright
PyYAML
langchain
openai